+++ CODEOWNERS
@@ -47,9 +47,6 @@ tests/scripts/t3000/ @tenstorrent/metalium-developers-infra
 tests/scripts/tg/ @tenstorrent/metalium-developers-infra
 tests/scripts/tgg/ @tenstorrent/metalium-developers-infra
 
-# fabric
-tt_fabric/ @ubcheema @aliuTT @aagarwalTT
-
 # Metalium - public API
 tt_metal/api @abhullar-tt @pgkeller @aliuTT @tt-aho @tt-dma @tt-asaigal @ubcheema @cfjchu @omilyutin-tt
 
@@ -58,6 +55,9 @@ tt_metal/ @abhullar-tt @pgkeller @aliuTT @tt-aho @tt-dma @tt-asaigal @ubcheema @
 tt_metal/host_api.hpp @abhullar-tt @pgkeller @aliuTT @tt-aho @tt-dma @tt-asaigal @davorchap @cfjchu @omilyutin-tt
 tt_metal/impl/device/ @abhullar-tt @pgkeller @aliuTT @tt-aho @tt-dma @tt-asaigal @ubcheema @davorchap @cfjchu @omilyutin-tt
 tt_metal/**/requirements*.txt @tenstorrent/metalium-developers-infra
+
+# fabric
+tt_metal/fabric/ @ubcheema @aliuTT @aagarwalTT
 
 # metal - dispatch
 tt_metal/impl/dispatch/kernels/packet_* @ubcheema @aliuTT
