+++ tests/tt_metal/tt_fabric/CMakeLists.txt
@@ -5,7 +5,7 @@ target_link_libraries(
     fabric_unit_tests
     PRIVATE
         tt_metal
-        tt_fabric
+        fabric
         test_common_libs
 )
 
@@ -13,7 +13,6 @@ target_include_directories(
     fabric_unit_tests
     PRIVATE
         ${UMD_HOME}
-        ${PROJECT_SOURCE_DIR}/tt_fabric
         ${PROJECT_SOURCE_DIR}/tests
         ${PROJECT_SOURCE_DIR}/tt_metal
         ${CMAKE_CURRENT_SOURCE_DIR}/common
