+++ tt_metal/fabric/CMakeLists.txt
@@ -1,29 +1,29 @@
-add_library(tt_fabric)
-add_library(TT::Fabric ALIAS tt_fabric)
+add_library(fabric)
+add_library(TT::Fabric ALIAS fabric)
 
 target_sources(
-    tt_fabric
+    fabric
     PRIVATE
         control_plane.cpp
         routing_table_generator.cpp
         mesh_graph.cpp
 )
 
-target_include_directories(tt_fabric PRIVATE .)
+target_include_directories(fabric PRIVATE .)
 
 target_link_libraries(
-    tt_fabric
+    fabric
     PRIVATE
-        Metalium::Metal
         umd::device
-        metal_common_libs
         magic_enum
         fmt::fmt-header-only
         yaml-cpp::yaml-cpp
+        Metalium::Metal::Impl
+        TT::Metalium::HostDevCommon
 )
 
 target_precompile_headers(
-    tt_fabric
+    fabric
     PRIVATE
         <functional>
         <map>
@@ -33,10 +33,10 @@ target_precompile_headers(
         <vector>
 )
 
-target_compile_options(tt_fabric PRIVATE -Wno-int-to-pointer-cast)
+target_compile_options(fabric PRIVATE -Wno-int-to-pointer-cast)
 
 set_target_properties(
-    tt_fabric
+    fabric
     PROPERTIES
         INSTALL_RPATH
             "${PROJECT_BINARY_DIR}/lib"
