+++ tt_metal/fabric/mesh_graph_descriptors/n300_mesh_graph_descriptor.yaml
