+++ tt_metal/fabric/mesh_graph_descriptors/quanta_galaxy_mesh_graph_descriptor.yaml
@@ -1,0 +1,27 @@
+ChipSpec: {
+  arch: wormhole_b0,
+  ethernet_ports: {
+    N: 4,
+    E: 4,
+    S: 4,
+    W: 4,
+  }
+}
+
+
+Board: [
+  { name: Galaxy,
+    type: Mesh,
+    topology: [4, 8]}
+ ]
+
+Mesh: [
+{
+  id: 0,
+  board: Galaxy,
+  topology: [1, 1],
+  host_mapping: [[]]}
+]
+
+Graph: [
+]
