+++ tt_metal/fabric/mesh_graph_descriptors/t3k_mesh_graph_descriptor.yaml
