+++ tt_metal/fabric/mesh_graph_descriptors/tg_mesh_graph_descriptor.yaml
